requirements.txt
altair
holoviews
hvplot
ipywidgets_bokeh
jupyter-bokeh
jupyter-panel-proxy==0.2.0a2
jupyter-server-proxy
jupyterlab
matplotlib
pandas
panel
plotly
plotnine
scipy
seaborn
vega_datasets
pydeck
pyecharts
vtk
pyvista
xarray
pooch
datashader
folium
pytest
ipysheet==0.4.4
ipydatagrid